[mod] streaks

+++ AtCoder/key2020/b.cpp
@@ -3,37 +3,37 @@
 #define rep1(i, n) for (int i = 1; i < (n); ++i)
 using namespace std;
 typedef long long ll;
-
+ 
 int main() {
     cin.tie(nullptr);
     ios::sync_with_stdio(false);
-
-    ll n;
+ 
+    int n;
     cin >> n;
-
-    pair<ll, ll> xl[n];
+ 
+    pair<int, int> xl[n];
     rep(i, n) {
-        ll x, l;
+        int x, l;
         cin >> x >> l;
-
+ 
         xl[i] = make_pair(x, l);
     }
     sort(xl, xl+n);
-
-    ll res = n;
-    ll p_reach = xl[0].first + xl[0].second;
+ 
+    int res = n;
+    int p_reach = xl[0].first + xl[0].second;
     rep1(i, n){
-        ll x = xl[i].first;
-        ll l = xl[i].second;
-
+        int x = xl[i].first;
+        int l = xl[i].second;
+ 
         if(x-l < p_reach) {
             --res;
-
+ 
         }
         else p_reach = x+l;
     }
     
     cout << res << endl;
-
+ 
     return 0;
 }
